fixed view

+++ src/systems/render_system.cpp
@@ -1107,17 +1107,17 @@ void RenderSystem::drawHUD(EntityManager& em, ENGI::GameEngine& engine, bool deb
             auto &aic = em.getComponent<AIComponent>(e);
             if(aic.alert_state){
                 auto& r = em.getComponent<RenderComponent>(e);
-                int barX = static_cast<int>(engine.getWorldToScreenX(r.position));
-                int barY = static_cast<int>(engine.getWorldToScreenY(r.position));
+                float barX = engine.getWorldToScreenX(r.position);
+                float barY = engine.getWorldToScreenY(r.position);
                 // Obtén las coordenadas del triángulo
-                Vector2 point1 = {barX, barY - 120};
-                Vector2 point2 = {barX - 30, barY - 50 };
-                Vector2 point3 = {barX + 30, barY - 50};
+                Vector2 point1 = {barX, barY - 120.0f};
+                Vector2 point2 = {barX - 30.0f, barY - 50.0f };
+                Vector2 point3 = {barX + 30.0f, barY - 50.0f};
         
                 // Dibuja el triángulo
                 DrawTriangle(point1, point2, point3, BLACK);
                 // Dibuja el signo de exclamación dentro del triángulo
-                engine.drawText("!", barX - 2, barY - 100, 50, YELLOW);
+                engine.drawText("!", static_cast<int>(barX - 2), static_cast<int>(barY - 100), 50, YELLOW);
             }
         }
